MAINTENANCE intelij auto code cleanup

+++ src/main/java/com/anthonykeane/speedzone/SeekBarPreference.java
@@ -125,7 +125,7 @@ public class SeekBarPreference extends Preference
     @Override
     protected Object onGetDefaultValue(TypedArray ta, int index) {
 
-        int dValue = (int) ta.getInt(index, 50);
+        int dValue = ta.getInt(index, 50);
 
         return validateValue(dValue);
     }
